update(cherryusb): update to v1.4.0

+++ components/drivers/usb/cherryusb/demo/cdc_acm_multi_template.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 #include "usbd_core.h"
-#include "usbd_cdc.h"
+#include "usbd_cdc_acm.h"
 
 /*!< endpoint address */
 #define CDC_IN_EP  0x81
@@ -223,7 +223,7 @@ struct usbd_interface intf5;
 struct usbd_interface intf6;
 struct usbd_interface intf7;
 
-void cdc_acm_multi_init(uint8_t busid, uint32_t reg_base)
+void cdc_acm_multi_init(uint8_t busid, uintptr_t reg_base)
 {
     usbd_desc_register(busid, cdc_descriptor);
