+++ components/drivers/usb/cherryusb/demo/hid_custom_inout_template.c
@@ -272,7 +272,7 @@ static struct usbd_endpoint custom_out_ep = {
  */
 struct usbd_interface intf0;
 
-void hid_custom_init(uint8_t busid, uint32_t reg_base)
+void hid_custom_init(uint8_t busid, uintptr_t reg_base)
 {
     usbd_desc_register(busid, hid_descriptor);
     usbd_add_interface(busid, usbd_hid_init_intf(busid, &intf0, hid_custom_report_desc, HID_CUSTOM_REPORT_DESC_SIZE));
